+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - (Bugfix) Unlock broken inspectors
 - (Debug) Allow to send package to stdout
 - (Improvement) ArangoDB image validation (=>3.10) for ARM64 architecture
+- (Improvement) Use inspector for ArangoMember
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress